requirements.txt
fastapi
python-dotenv
uvicorn[standard]
jinja2
python-multipart
pygments
requests
